fix: cmd/agent объединить env.go и flag.go = config.go
